Move clones to hierarcy package
