+++ core/src/main/scala/chisel3/Module.scala
@@ -200,146 +200,10 @@ package experimental {
 }
 
 package internal {
-  import chisel3.experimental.BaseModule
-  import chisel3.experimental.hierarchy.core.{Clone, IsInstantiable, Proto}
 
   object BaseModule {
 
-    /** Represents a clone of an underlying object. This is used to support CloneModuleAsRecord and Instance/Definition.
-      *
-      * @note We don't actually "clone" anything in the traditional sense but is a placeholder so we lazily clone internal state
-      */
-    trait IsClone[+T] {
-      // Underlying object of which this is a clone of
-      private[chisel3] def getProto: T
-
-      /** Determines whether another object is a clone of the same underlying proto
-        *
-        * @param a
-        */
-      def hasSameProto(a: Any): Boolean = {
-        val aProto = a match {
-          case x: IsClone[BaseModule] => x.getProto
-          case o => o
-        }
-        this == aProto || getProto == aProto
-      }
-    }
-
-    // Private internal class to serve as a _parent for Data in cloned ports
-    private[chisel3] class ModuleClone[T <: BaseModule](val getProto: T) extends PseudoModule with IsClone[T] {
-      override def toString = s"ModuleClone(${getProto})"
-      // Do not call default addId function, which may modify a module that is already "closed"
-      override def addId(d: HasId): Unit = ()
-      def getPorts = _portsRecord
-      // ClonePorts that hold the bound ports for this module
-      // Used for setting the refs of both this module and the Record
-      private[BaseModule] var _portsRecord: Record = _
-      // This is necessary for correctly supporting .toTarget on a Module Clone. If it is made from the
-      // Instance/Definition API, it should return an instanceTarget. If made from CMAR, it should return a
-      // ModuleTarget.
-      private[chisel3] var _madeFromDefinition: Boolean = false
-      // Don't generate a component, but point to the one for the cloned Module
-      private[chisel3] def generateComponent(): Option[Component] = {
-        require(!_closed, "Can't generate module more than once")
-        _closed = true
-        _component = getProto._component
-        None
-      }
-      // Maps proto ports to module clone's ports
-      private[chisel3] lazy val ioMap: Map[Data, Data] = {
-        getProto match {
-          // BlackBox needs special handling for its pseduo-io Bundle
-          case protoBB: BlackBox =>
-            Map(protoBB._io.get -> getPorts.elements("io"))
-          case _ =>
-            val name2Port = getPorts.elements
-            getProto.getChiselPorts.map { case (name, data) => data -> name2Port(name) }.toMap
-        }
-      }
-      // This module doesn't actually exist in the FIRRTL so no initialization to do
-      private[chisel3] def initializeInParent(parentCompileOptions: CompileOptions): Unit = ()
-
-      // Name of this instance's module is the same as the proto's name
-      override def desiredName: String = getProto.name
-
-      private[chisel3] def setRefAndPortsRef(namespace: Namespace): Unit = {
-        val record = _portsRecord
-        // Use .forceName to re-use default name resolving behavior
-        record.forceName(None, default = this.desiredName, namespace)
-        // Now take the Ref that forceName set and convert it to the correct Arg
-        val instName = record.getRef match {
-          case Ref(name) => name
-          case bad       => throwException(s"Internal Error! Cloned-module Record $record has unexpected ref $bad")
-        }
-        // Set both the record and the module to have the same instance name
-        val ref = ModuleCloneIO(getProto, instName)
-        record.setRef(ref, force = true) // force because we did .forceName first
-        getProto match {
-          // BlackBox needs special handling for its pseduo-io Bundle
-          case _: BlackBox =>
-            // Override the io Bundle's ref so that it thinks it is the top for purposes of
-            // generating FIRRTL
-            record.elements("io").setRef(ref, force = true)
-          case _ => // Do nothing
-        }
-
-        this.setRef(Ref(instName))
-      }
-    }
-
-    /** Represents a module viewed from a different instance context.
-      *
-      * @note Why do we need both ModuleClone and InstanceClone? If we are annotating a reference in a module-clone,
-      * all submodules must be also be 'cloned' so the toTarget can be computed properly. However, we don't need separate
-      * connectable ports for this instance; all that's different from the proto is the parent.
-      *
-      * @note In addition, the instance name of an InstanceClone is going to be the SAME as the proto, but this is not true
-      * for ModuleClone.
-      */
-    private[chisel3] final class InstanceClone[T <: BaseModule](val getProto: T, val instName: () => String)
-        extends PseudoModule
-        with IsClone[T] {
-      override def toString = s"InstanceClone(${getProto})"
-      // No addition components are generated
-      private[chisel3] def generateComponent(): Option[Component] = None
-      // Necessary for toTarget to work
-      private[chisel3] def setAsInstanceRef(): Unit = { this.setRef(Ref(instName())) }
-      // This module doesn't acutally exist in the FIRRTL so no initialization to do
-      private[chisel3] def initializeInParent(parentCompileOptions: CompileOptions): Unit = ()
-      // Instance name is the same as proto's instance name
-      override def instanceName = instName()
-      // Module name is the same as proto's module name
-      override def desiredName: String = getProto.name
-    }
-
-    /** Represents a Definition root module, when accessing something from a definition
-      *
-      * @note This is necessary to distinguish between the toTarget behavior for a Module returned from a Definition,
-      * versus a normal Module. A normal Module.toTarget will always return a local target. If calling toTarget
-      * on a Module returned from a Definition (and thus wrapped in an Instance), we need to return the non-local
-      * target whose root is the Definition. This DefinitionClone is used to represent the root parent of the
-      * InstanceClone (which represents the returned module).
-      */
-    private[chisel3] class DefinitionClone[T <: BaseModule](val getProto: T) extends PseudoModule with IsClone[T] {
-      override def toString = s"DefinitionClone(${getProto})"
-      // No addition components are generated
-      private[chisel3] def generateComponent(): Option[Component] = None
-      // Do not call default addId function, which may modify a module that is already "closed"
-      override def addId(d: HasId): Unit = ()
-      // Necessary for toTarget to work
-      private[chisel3] def initializeInParent(parentCompileOptions: CompileOptions): Unit = ()
-      // Module name is the same as proto's module name
-      override def desiredName: String = getProto.name
-    }
-
-    /** @note If we are cloning a non-module, we need another object which has the proper _parent set!
-      */
-    trait InstantiableClone[T <: IsInstantiable] extends IsClone[T] {
-      private[chisel3] def _innerContext: experimental.hierarchy.Hierarchy[_]
-      private[chisel3] def getInnerContext: Option[BaseModule] = _innerContext.getInnerDataContext
-    }
-
+    import chisel3.experimental.hierarchy._
     /** Record type returned by CloneModuleAsRecord
       *
       * @note These are not true Data (the Record doesn't correspond to anything in the emitted
@@ -545,9 +409,9 @@ package experimental {
       * @note Should not be called until circuit elaboration is complete
       */
     final def toTarget: ModuleTarget = this match {
-      case m: internal.BaseModule.InstanceClone[_] =>
+      case m: experimental.hierarchy.InstanceClone[_] =>
         throwException(s"Internal Error! It's not legal to call .toTarget on an InstanceClone. $m")
-      case m: internal.BaseModule.DefinitionClone[_] =>
+      case m: experimental.hierarchy.DefinitionClone[_] =>
         throwException(s"Internal Error! It's not legal to call .toTarget on an DefinitionClone. $m")
       case _ => ModuleTarget(this.circuitName, this.name)
     }
@@ -564,12 +428,12 @@ package experimental {
       * the correct [[InstanceTarget]]s whenever using the Definition/Instance API.
       */
     private[chisel3] def getTarget: IsModule = this match {
-      case m: internal.BaseModule.InstanceClone[_] if m._parent.nonEmpty =>
+      case m: experimental.hierarchy.InstanceClone[_] if m._parent.nonEmpty =>
         m._parent.get.getTarget.instOf(instanceName, name)
-      case m: internal.BaseModule.ModuleClone[_] if m._madeFromDefinition =>
+      case m: experimental.hierarchy.ModuleClone[_] if m._madeFromDefinition =>
         m._parent.get.getTarget.instOf(instanceName, name)
       // Without this, we get the wrong CircuitName for the Definition
-      case m: internal.BaseModule.DefinitionClone[_] if m._circuit.nonEmpty =>
+      case m: experimental.hierarchy.DefinitionClone[_] if m._circuit.nonEmpty =>
         ModuleTarget(this._circuit.get.circuitName, this.name)
       case _ => this.toTarget
     }
