+++ core/src/main/scala/chisel3/RawModule.scala
@@ -8,7 +8,7 @@ import scala.language.experimental.macros
 import scala.annotation.nowarn
 import chisel3.experimental.BaseModule
 import chisel3.internal._
-import chisel3.internal.BaseModule.{InstanceClone, ModuleClone}
+import chisel3.experimental.hierarchy.{InstanceClone, ModuleClone}
 import chisel3.internal.Builder._
 import chisel3.internal.firrtl._
 import chisel3.internal.sourceinfo.UnlocatableSourceInfo
