+++ src/main/scala/chisel3/aop/Select.scala
@@ -9,7 +9,7 @@ import chisel3.experimental.FixedPoint
 import chisel3.internal.firrtl.{Definition => DefinitionIR, _}
 import chisel3.experimental.hierarchy.core._
 import chisel3.internal.PseudoModule
-import chisel3.internal.BaseModule.ModuleClone
+import chisel3.experimental.hierarchy.ModuleClone
 import firrtl.annotations.ReferenceTarget
 import scala.reflect.runtime.universe.TypeTag
 
@@ -53,7 +53,7 @@ object Select {
         d.commands.collect {
           case d: DefInstance =>
             d.id match {
-              case p: chisel3.internal.BaseModule.IsClone[_] =>
+              case p: IsClone[_] =>
                 parent._lookup { x => new Instance(Clone(p)).asInstanceOf[Instance[BaseModule]] }
               case other: BaseModule =>
                 parent._lookup { x => other }
@@ -78,7 +78,7 @@ object Select {
         d.commands.flatMap {
           case d: DefInstance =>
             d.id match {
-              case p: chisel3.internal.BaseModule.IsClone[_] =>
+              case p: IsClone[_] =>
                 val i = parent._lookup { x => new Instance(Clone(p)).asInstanceOf[Instance[BaseModule]] }
                 if (i.isA[T]) Some(i.asInstanceOf[Instance[T]]) else None
               case other: BaseModule =>
@@ -117,7 +117,7 @@ object Select {
         d.commands.collect {
           case i: DefInstance =>
             i.id match {
-              case p: chisel3.internal.BaseModule.IsClone[_] =>
+              case p: IsClone[_] =>
                 parent._lookup { x => new Definition(Proto(p.getProto)).asInstanceOf[Definition[BaseModule]] }
               case other: BaseModule =>
                 parent._lookup { x => other.toDefinition }
@@ -148,7 +148,7 @@ object Select {
         d.commands.flatMap {
           case d: DefInstance =>
             d.id match {
-              case p: chisel3.internal.BaseModule.IsClone[_] =>
+              case p: IsClone[_] =>
                 val d = parent._lookup { x => new Definition(Clone(p)).asInstanceOf[Definition[BaseModule]] }
                 if (d.isA[T]) Some(d.asInstanceOf[Definition[T]]) else None
               case other: BaseModule =>
